requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
pillow==10.2.0
sqlparse==0.4.4
